Attach map input gesture directly
