+++ Ascension/Modules/Arkheion/Core/ArkheionMapView+HitTesting.swift
@@ -102,4 +102,25 @@ extension ArkheionMapView {
         let proj = CGPoint(x: start.x + t * dx, y: start.y + t * dy)
         return hypot(point.x - proj.x, point.y - proj.y)
     }
+
+    /// Returns true if the canvas point lies within any node's hit radius.
+    func nodeHitCheck(at point: CGPoint, in geo: GeometryProxy) -> Bool {
+        let center = CGPoint(x: geo.size.width / 2, y: geo.size.height / 2)
+        for branch in store.branches {
+            guard let ring = store.rings.first(where: { $0.ringIndex == branch.ringIndex }) else { continue }
+            let direction = CGPoint(x: cos(branch.angle), y: sin(branch.angle))
+            for (i, node) in branch.nodes.enumerated() {
+                let distance = ring.radius + CGFloat(i + 1) * 60
+                let nodePos = CGPoint(
+                    x: center.x + direction.x * distance,
+                    y: center.y + direction.y * distance
+                )
+                let radius = max(node.size.radius, 12) + 10
+                if hypot(point.x - nodePos.x, point.y - nodePos.y) <= radius {
+                    return true
+                }
+            }
+        }
+        return false
+    }
 }
